allow a share to be updated
